+++ scripts/alternativeAdvantage.js
@@ -1,6 +1,6 @@
 export function setupAlternativeAdvantage() {
   const libWrapper = globalThis.libWrapper;
-	if (libWrapper) libWrapper.register("dnd5e-scriptlets", "CONFIG.Dice.D20Roll.prototype.configureModifiers", configureAdvantage, "WRAPPER");
+  if (libWrapper) libWrapper.register("dnd5e-scriptlets", "CONFIG.Dice.D20Roll.prototype.configureModifiers", configureAdvantage, "WRAPPER");
 
   /*
   Hooks.on("dnd5e.preRollAttack", (...args) => {
@@ -15,7 +15,7 @@ export function checkAlternativeAdvatage() {
   if (!globalThis.libWrapper) {
     const errorMessage = game.i18n.localize("dnd5e-scriptlets.needsLibWrapper");
     console.error(errorMessage);
-    ui.notifications?.error(errorMessage, {permanent: true});
+    ui.notifications?.error(errorMessage, { permanent: true });
   }
 }
 
@@ -26,17 +26,27 @@ function configureAdvantage(wrapped) {
   const d20 = this.terms[0];
   if (this.hasAdvantage) {
     d20.modifiers = d20.modifiers?.filter(s => s !== "kh");
-    this.terms.push(new OperatorTerm({operator: "+"}));
-    const extraTerms = new Roll(advantageQuantum, this.data).terms;
-    this.terms = this.terms.concat(extraTerms)
     d20.number = 1;
-  }
+    if (!d20.options.hasAdvantage) {
+      this.terms.push(new OperatorTerm({ operator: "+", options: { isAdvantageTerm: true } }));
+      const extraTerms = new Roll(advantageQuantum, this.data).terms;
+      for (let term of extraTerms) term.options.isAdvantageTerm = true;
+      this.terms = this.terms.concat(extraTerms)
+      d20.options.hasAdvantage = true;
+    }
+  } else this.terms = this.terms.filter(t => !t.options.isAdvantageTerm);
+
   if (this.hasDisadvantage) {
     d20.modifiers = d20.modifiers?.filter(s => s !== "kl");
-    this.terms.push(new OperatorTerm({operator: "-"}));
-    const extraTerms = new Roll(advantageQuantum, this.data).terms;
-    this.terms = this.terms.concat(extraTerms)
     d20.number = 1;
-  }
+    if (!d20.options.hasDisadvantage) {
+      this.terms.push(new OperatorTerm({ operator: "-", options: { isDisadvantageTerm: true } }));
+      const extraTerms = new Roll(advantageQuantum, this.data).terms;
+      for (let term of extraTerms) term.options.isDisadvantageTerm = true;
+      this.terms = this.terms.concat(extraTerms)
+      d20.options.hasDisadvantage = true;
+    }
+  } else this.terms = this.terms.filter(t => !t.options.isDisadvantageTerm);
+
   this._formula = this.constructor.getFormula(this.terms);
 }